Add support for DetectorState on Dashboard. Issue: #25
