+++ public/pages/DetectorsList/utils/tableUtils.tsx
@@ -19,8 +19,8 @@ import moment from 'moment';
 import get from 'lodash/get';
 import React from 'react';
 import { Detector } from '../../../../server/models/types';
-import { PLUGIN_NAME } from '../../../utils/constants';
-import { DETECTOR_STATE, stateToColorMap } from '../../utils/constants';
+import { PLUGIN_NAME, DETECTOR_STATE } from '../../../utils/constants';
+import { stateToColorMap } from '../../utils/constants';
 
 export const DEFAULT_EMPTY_DATA = '-';
